pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-signal-platform"
version = "0.1.0"
description = "Production trading signal platform with Telegram bot integration"
requires-python = ">=3.11"
authors = [
    {name = "Trading Signals Team", email = "dev@tradingsignals.local"},
]
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.4.2",
    "pydantic-settings>=2.0.3",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg[binary]>=3.17.0",
    "aiosqlite>=3.14.0",
    "redis>=5.0.0",
    "httpx>=0.25.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.1",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "cryptography>=41.0.7",
    "pyopenssl>=23.3.0",
    "orjson>=3.9.10",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "isort>=5.13.2",
    "pre-commit>=3.5.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [".venv/*", "venv/*"]
known_first_party = ["backend"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "W191",   # indentation contains tabs
    "E401",   # multiple imports on one line (conflicts with isort)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["backend/tests"]
asyncio_mode = "auto"
addopts = "--strict-markers --tb=short"
markers = [
    "asyncio: marks tests as async",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
